requirements.txt
anyascii==0.3.0
asgiref==3.5.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.50.0
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==5.3
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fonttools==4.31.2
html5lib==1.1
idna==3.3
kiwisolver==1.4.2
l18n==2021.3
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
telepath==0.2
urllib3==1.26.9
uuid==1.30
wagtail==2.15.4
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
wagtailmedia
